\name{mcConsensus-package}
\alias{mcConsensus-package}
\alias{mcConsensus}
\docType{package}
\title{
Multicore consensus clustering.
}
\description{
Package contains methods for performing heavy consensus clustering algorithm using multicore processors. This package was based on clusterCons package.
}
\details{
\tabular{ll}{
Package: \tab mcConsensus\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-07-19\cr
License: \tab GPL-3\cr
}
}
\author{
Tadeusz Tagowski

Maintainer: Tadeusz Tagowski <tadtagov@gmail.com>
}
\references{
Stefano Monti, Pablo Tamayo, Jill Mesirov, and Todd Golub. Consensus clustering: a
resampling-based method for class discovery and visualization of gene expression microar-
ray data. Machine learning, 52(1-2):91–118, 2003.

T Ian Simpson, J Douglas Armstrong, and Andrew Jarman. Merged consensus clustering
to assess and improve class discovery with microarray data. BMC bioinformatics, 11(1):590,
2010.

}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
